+++ service/controller/resource/awsclient/create.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/giantswarm/microerror"
-	"github.com/giantswarm/operatorkit/v2/pkg/controller/context/reconciliationcanceledcontext"
+	"github.com/giantswarm/operatorkit/v4/pkg/controller/context/reconciliationcanceledcontext"
 )
 
 func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {