+++ service/internal/asg/asg.go
@@ -3,11 +3,12 @@ package asg
 import (
 	"context"
 	"fmt"
+	"sort"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/autoscaling"
 	"github.com/aws/aws-sdk-go/service/ec2"
-	"github.com/giantswarm/k8sclient/v4/pkg/k8sclient"
+	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/to"
 	"k8s.io/apimachinery/pkg/api/meta"
@@ -87,18 +88,7 @@ func (a *ASG) Drainable(ctx context.Context, obj interface{}) (string, error) {
 	// interested it.
 	var asgs []*autoscaling.Group
 	{
-		var names []string
-		{
-			m := map[string]struct{}{}
-
-			for _, i := range instances {
-				m[asgNameFromInstance(i)] = struct{}{}
-			}
-
-			for k := range m {
-				names = append(names, k)
-			}
-		}
+		names := namesFromInstances(instances)
 
 		asgs, err = a.cachedASGs(ctx, cr, names)
 		if err != nil {
@@ -179,6 +169,10 @@ func (a *ASG) lookupASGs(ctx context.Context, names []string) ([]*autoscaling.Gr
 	cc, err := controllercontext.FromContext(ctx)
 	if err != nil {
 		return nil, microerror.Mask(err)
+	}
+
+	if len(names) == 0 {
+		return nil, microerror.Mask(noASGError)
 	}
 
 	var asgs []*autoscaling.Group
@@ -275,6 +269,34 @@ func drainable(ctx context.Context, asgs []*autoscaling.Group) (string, error) {
 	}
 
 	return "", microerror.Mask(noDrainableError)
+}
+
+func namesFromInstances(instances []*ec2.Instance) []string {
+	var names []string
+	{
+		m := map[string]struct{}{}
+
+		for _, i := range instances {
+			// While EC2 instances are transitioning due to scale in and scale
+			// out events they may not be associated with any ASG. In these
+			// cases we ignore the instances and move on to the next one.
+			// Eventually we will collect the relevant ASG names.
+			n := asgNameFromInstance(i)
+			if n == "" {
+				continue
+			}
+
+			m[n] = struct{}{}
+		}
+
+		for k := range m {
+			names = append(names, k)
+		}
+	}
+
+	sort.Strings(names)
+
+	return names
 }
 
 func toPtrList(l []string) []*string {